% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleInterest.R
\name{amount_si}
\alias{amount_si}
\title{Calculate Amount for Simple Interest}
\usage{
amount_si(principle, rate, tp = 1)
}
\arguments{
\item{principle}{double}

\item{rate}{double}

\item{tp}{double 1}
}
\value{
double
}
\description{
Takes principle, rate of interest(per annum) and time(in years) as input
and returns the amount.
If time period is not supplied default value of 1 is used.
}
\examples{
amount_si(5000, 3, 4)

p <- 3000
r <- 2.5
t <- 10
amount_si(p, r, t)

p <- 3000
r <- 6
amount_si(p, r)
}
