% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_interest.R
\name{compound_interest}
\alias{compound_interest}
\title{calculate Compound Interest}
\usage{
compound_interest(p, r, t)
}
\arguments{
\item{p}{numeric}

\item{r}{numeric}

\item{t}{numeric}
}
\value{
numeric
}
\description{
calculate Compound Interest
}
\examples{
compound_interest(1000, 0.0425, 6)
}
