% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleInterest.R
\name{si}
\alias{si}
\title{Calculate Simple Interest}
\usage{
si(principle, rate, tp = 1)
}
\arguments{
\item{principle}{double}

\item{rate}{double}

\item{tp}{double 1}
}
\value{
double
}
\description{
Takes principle, rate of interest(per annum) and time(in years) as input
and returns the simple interest.
If time period is not supplied default value of 1 is used.
}
\examples{
si(5000, 3, 4)

p <- 3000
r <- 2.5
t <- 10
si(p, r, t)

p <- 3000
r <- 6
si(p, r)
}
