% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_interest.R
\name{simple_interest}
\alias{simple_interest}
\title{Calculate Simple Interest}
\usage{
simple_interest(p, r, t)
}
\arguments{
\item{p}{numeric}

\item{r}{numeric}

\item{t}{numeric}
}
\value{
numeric
}
\description{
Calculate Simple Interest
}
\examples{
simple_interest(5000,0.08,11/12) #366.67
}
